refactor(architecture): move version into release domain
