+++ src/lib.rs
@@ -20,4 +20,5 @@ pub mod utils;
 // Re-export everything from core for ergonomic library use
 // Users can write `homeboy::config` instead of `homeboy::core::config`
 pub use core::*;
+pub use core::release::version;
 pub use utils::*;
